backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & JWT
PyJWT==2.8.0
cryptography==42.0.2

# API & Serialization
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.1

# Environment
python-decouple==3.8
python-dotenv==1.0.1

# External APIs
requests==2.31.0
httpx==0.26.0

# Utilities
celery==5.3.6
redis==5.0.1

# Development & Testing
pytest==8.0.0
pytest-django==4.8.0
pytest-cov==4.1.0
factory-boy==3.3.0
freezegun==1.4.0

# Code Quality
flake8==7.0.0
black==24.1.1
isort==5.13.2

# Production
gunicorn==21.2.0
whitenoise==6.6.0